+++ components/CommentBox.tsx
@@ -1,7 +1,7 @@
 import { MarkdownIcon, MarkGithubIcon, SignOutIcon, TypographyIcon } from '@primer/octicons-react';
 import { ChangeEvent, useCallback, useContext, useEffect, useRef, useState } from 'react';
 import { adaptComment, adaptReply, handleCommentClick, processCommentBody } from '../lib/adapter';
-import { AuthContext } from '../lib/context';
+import { AuthContext, DialogContext } from '../lib/context';
 import { useGiscusTranslation } from '../lib/i18n';
 import { IComment, IReply } from '../lib/types/adapter';
 import { resizeTextArea } from '../lib/utils';
@@ -39,6 +39,7 @@ export default function CommentBox({
   const [isFixedWidth, setIsFixedWidth] = useState(false);
   const [lastHeight, setLastHeight] = useState('');
   const { token, origin, getLoginUrl, onSignOut } = useContext(AuthContext);
+  const { alert } = useContext(DialogContext);
   const textarea = useRef<HTMLTextAreaElement>(null);
   const loginUrl = getLoginUrl(origin);
   const isReply = !!replyToId;
@@ -74,7 +75,7 @@ export default function CommentBox({
     const payload = { body: input, discussionId: id, replyToId };
 
     if (!id) {
-      window.alert('Unable to create discussion.');
+      await alert({ message: t('genericError', { message: 'Unable to create discussion.' }) });
       setIsSubmitting(false);
       return;
     }
@@ -103,6 +104,8 @@ export default function CommentBox({
     replyToId,
     onDiscussionCreateRequest,
     token,
+    alert,
+    t,
     onSubmit,
     reset,
   ]);